For Issue #155: refactored macroeditor and communication + fixed bug
